pyproject.toml
[build-system]
requires = ["setuptools>=42,<43", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-stack"
version = "0.1.0"
description = "A modern, maintainable, and performant Python implementation of the LOCAL-LLM-STACK"
readme = "README.md"
authors = [
    {name = "Data Mint Research", email = "info@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0,<9.0.0",
    "docker>=6.0.0,<7.0.0",
    "pyyaml>=6.0,<7.0",
    "rich>=12.0.0,<13.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "neo4j>=5.0.0,<6.0.0",
    "jinja2>=3.0.0,<4.0.0",
    "requests>=2.0.0,<3.0.0",
    "pyupgrade>=3.0.0,<4.0.0",
    "isort>=5.0.0,<6.0.0",
    "black>=23.0.0,<24.0.0",
    "vulture>=2.0.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<8.0.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "black>=23.0.0,<24.0.0",
    "isort>=5.0.0,<6.0.0",
    "mypy>=1.0.0,<2.0.0",
    "flake8>=6.0.0,<7.0.0",
]

[project.scripts]
llm = "llm_stack.cli:main"

[tool.setuptools]
packages = ["llm_stack"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"